refactor: move tab to components directory
